add ejs tests
