--- packages/web3/test/black_box/jest.config.js
@@ -1,5 +1,0 @@
-/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
-module.exports = {
-	preset: 'ts-jest',
-	testEnvironment: 'node',
-};
