+++ packages/web3/test/cjs_black_box/cjs_tests/web3-eth-accounts/hashMessage.test.ts
@@ -20,7 +20,7 @@ import {
 	closeOpenConnection,
 	getSystemTestProvider,
 	isWs,
-} from '../../fixtures/system_tests_utils';
+} from '../../../esm_black_box/test/fixtures/system_tests_utils';
 /* eslint-disable @typescript-eslint/no-unsafe-call */
 // eslint-disable-next-line @typescript-eslint/no-require-imports, @typescript-eslint/no-var-requires
 const Web3 = require('web3').default;
