+++ packages/web3/test/esm_black_box/jest.config.js
@@ -1,0 +1,9 @@
+/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
+export default {
+	'ts-jest': {
+		tsconfig: '../config/tsconfig.esm.json'
+	},
+	"transform": {
+		"^.+\\.(ts|tsx)$": "ts-jest"
+	  },
+};
