+++ packages/web3/test/esm_black_box/test/fixtures/contracts/ERC20Token.ts
