+++ packages/web3/test/esm_black_box/test/fixtures/contracts/ERC721Token.ts
