+++ packages/web3/test/esm_black_box/tsconfig.base.json
@@ -1,0 +1,20 @@
+{
+	"compilerOptions": {
+		"resolveJsonModule": true,
+		"forceConsistentCasingInFileNames": true,
+		"target": "es2016",
+		"moduleResolution": "node",
+		"newLine": "lf",
+		"noFallthroughCasesInSwitch": true,
+		"noImplicitReturns": true,
+		"noUnusedLocals": true,
+		"noUnusedParameters": true,
+		"pretty": true,
+		"removeComments": false,
+		"sourceMap": true,
+		"strict": true,
+		"strictNullChecks": true,
+		"esModuleInterop": true,
+		"allowSyntheticDefaultImports": true
+	}
+}
